Add files via upload
By referring to the principle of neuron side inhibition, a direct but effective preprocessing algorithm emphasizing fundamental frequency and harmonics (EFF)  is provided. I embed it into DNN algorithm or SNMF algorithm for speech enhancement, and find that EFF can significantly improve the perceptionality of speech, that is, improve the PESQ score, but reduce the intelligibility, that is, the STOI score will decrease.